Fix pets array scaffolding and clean up auth
